% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comicR.R
\name{renderComicR}
\alias{renderComicR}
\title{Widget render function for use in Shiny}
\usage{
renderComicR(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

