% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/comicR.R
\name{comicR}
\alias{comicR}
\title{Add a comic effect to R plots and other outputs}
\usage{
comicR(selector = NULL, ff = NULL, ffc = NULL, fsteps = NULL,
  msteps = NULL, width = 0, height = 0)
}
\description{
<Add Description>
}

